Fix potential false positive hydra results
In the previous implementation we would set `data["vulnerable"] = false` if hydra exited with an error.
Now it returns None in the error case.

This might still not be totally correct. For example ssh errors out when the server does not support password login. In this case the server is not vulnerable.

It does not appear that error messages are standardized across different ssh servers, so it is not trivial to handle this case.

For now we will err on the side of caution.